Move mainchain registration to common file
